+++ src/rules/alias/__test__/project/standalone.ts
@@ -1,0 +1,1 @@
+export const standalone = 1;